Refactored classes into header and implementation paired files.
